+++ src/css/0-site/index.scss
@@ -41,7 +41,6 @@ body {
     display: flex;
     height: 100%;
     padding: 0;
-    overflow: hidden;
 }
   
 h1 {
