+++ src/css/1-pages/dashboard.scss
@@ -7,8 +7,8 @@
 
     &-projects {
         display: grid;
-        grid-template-columns: repeat(6, 1fr);
-        gap: 12px;
+        grid-template-columns: repeat(7, 1fr);
+        gap: 24px;
     }
 
     &-separator {
