+++ src/css/2-components/navbar.scss
@@ -15,6 +15,9 @@
     &-hld {
         display: flex;
         background-color: variables.$c-beige;
+        position: sticky;
+        top: 0;
+        z-index: 100;
         .nav {
             &__btn {
                 margin-left: 10px;
