+++ src/types/ProjectData.ts
@@ -1,0 +1,6 @@
+import AddProjectData from './AddProjectData';
+
+export default interface ProjectData extends AddProjectData {
+  id: string;
+  lastUpdate: string;
+}
